requirements.txt
speechrecognition
numpy
matplotlib
opencv-python
easygui
Pillow
